Minor import clean-up and renaming
