+++ tests/types/cross_module_resolvers/a_mod.py
