+++ tests/types/cross_module_resolvers/b_mod.py
