+++ tests/types/cross_module_resolvers/c_mod.py
@@ -1,10 +1,12 @@
 from typing import List
 
-import strawberry
+import a_mod
+import b_mod
+import x_mod
+from a_mod import AObject as C_AObject
+from b_mod import BObject as C_BObject
 
-from . import a_mod, b_mod, x_mod
-from .a_mod import AObject as C_AObject
-from .b_mod import BObject as C_BObject
+import strawberry
 
 
 def c_inheritance_resolver() -> List["CInheritance"]:
